fixed departments and contacts
