+++ src/app/contacts/page.tsx
@@ -1,4 +1,4 @@
-import { DepartmentsKitPage } from "@/ui-pages/Departments/Kit/kit.pages";
+import { DepartmentsKitPage } from "@/ui-pages/Departments/Kit/kit.page";
 import { ReactElement } from "react";
 import {ContactsPage} from "@/ui-pages";
