+++ src/ui-pages/index.ts
@@ -1,3 +1,3 @@
 export { HomePage } from "./Home/home.page";
-export {DepartmentsKitPage} from "./Departments/Kit/kit.pages"
-export {ContactsPage} from "./Contacts/contacts.pages"
+export {DepartmentsKitPage} from "./Departments/Kit/kit.page"
+export {ContactsPage} from "./Contacts/contacts.page"
