+++ src/widgets/Department/AboutDepartments/aboutdepartments.module.css
