+++ src/widgets/Department/DepartmentSlider/DepartmentsSlider.module.scss
