+++ src/widgets/Department/index.tsx
@@ -1,0 +1,4 @@
+export { AboutDepartments } from "./AboutDepartments/AboutDepartments";
+export { DepartmentsTeachers } from "./DepartmentsTeachers/DepartmentsTeachers";
+export { DepartmentsTeachersCard } from "./DepartmentsTeachers/DepartmentsTeachersCard";
+export { DepartmentSlider } from "./DepartmentSlider/DepartmentSlider";
