+++ src/widgets/index.ts
@@ -7,10 +7,6 @@ export { LeedFormWithMap } from "./LeedFormWithMap/LeedFormWithMap";
 export { Intro } from "./Intro/Intro";
 export { Partners } from "./Partners/Partners";
 export { Stats } from "./Stats/Stats";
-export { DepartmentsCard } from "./DepartmentsCard/DepartmentsCard";
-export { AboutDepartments } from "./AboutDepartments/AboutDepartments";
-export { DepartmentsTeachers } from "./DepartmentsTeachers/DepartmentsTeachers";
-export { DepartmentsTeachersCard } from "./DepartmentsTeachers/DepartmentsTeachersCard";
-export { DepartmentSlider } from "@/widgets/DepartmentSlider/DepartmentSlider";
-export { Contacts } from "@/widgets/Contacts/Contacts";
+export * from "./Department";
+export { Contacts } from "./Contacts/Contacts";
